# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /course/weeks/{weekNumber}:
    get:
      operationId: getWeek
      tags: [course]
      summary: Get one week's lectures and assignment list
      parameters:
        - name: weekNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /course/lectures/{lectureId}:
    get:
      operationId: getLecture
      tags: [course]
      summary: Get the full lecture text for a topic
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
...
  /course/lectures/{lectureId}/expand:
    post:
      operationId: expandLecture
      tags: [course]
      summary: Generate the medium or long version of a single lecture on the spot
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureExpandInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
...
components:
  schemas:
...
    Lecture:
      type: object
      required: [id, topicId, title, weekNumber, body]
      properties:
        id: { type: integer }
        topicId: { type: integer }
        title: { type: string }
        weekNumber: { type: integer }
        body:
          type: string
          description: Short Markdown lecture text (the baseline / minimum-detail version). The frontend lets users select passages and send them to the tutor.
        bodyMedium:
          type: ["string", "null"]
          description: Medium-length version with more explanation and more examples. Null if not yet generated.
        bodyLong:
          type: ["string", "null"]
          description: Long version with the most explanation and the most examples. Null if not yet generated.
    Week:
      type: object
      required: [weekNumber, title, lectures, assignments]
      properties:
        weekNumber: { type: integer }
        title: { type: string }
        summary: { type: ["string", "null"] }
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/LectureRef"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentSummary"
    LectureRef:
      type: object
      required: [id, title, topicId]
      properties:
        id: { type: integer }
        title: { type: string }
        topicId: { type: integer }
    AssignmentSummary:
      type: object
      required: [id, kind, title, weekNumber, problemCount, isTimed, status]
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        title: { type: string }
        weekNumber: { type: integer }
        problemCount: { type: integer }
        isTimed: { type: boolean }
        timeLimitMinutes: { type: ["integer", "null"] }
        status:
          type: string
          enum: [not_started, in_progress, submitted]
        bestScore: { type: ["number", "null"] }
        lastAttemptId: { type: ["integer", "null"] }
...
    LectureExpandInput:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [medium, long]
...