# lib/api-zod/src/generated/types/lectureExpandInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LectureExpandInputLevel } from './lectureExpandInputLevel';

export interface LectureExpandInput {
  level: LectureExpandInputLevel;
}